--- frontend/src/metabase/timelines/questions/components/EventList/EventList.styled.tsx
@@ -1,5 +1,0 @@
-import styled from "@emotion/styled";
-
-export const ListRoot = styled.div`
-  padding: 1rem 0;
-`;
