--- frontend/src/metabase/timelines/questions/components/EventList/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./EventList";
