+++ frontend/src/metabase/timelines/questions/components/TimelineCard/TimelineCard.tsx
@@ -1,32 +1,30 @@
 import React, { ChangeEvent, memo, useCallback } from "react";
+import _ from "underscore";
+import { parseTimestamp } from "metabase/lib/time";
 import CheckBox from "metabase/core/components/CheckBox";
 import CollapseSection from "metabase/components/CollapseSection";
-import { Timeline } from "metabase-types/api";
-import EventList from "../EventList";
-import { CardHeader, CardTitle } from "./TimelineCard.styled";
+import { Timeline, TimelineEvent } from "metabase-types/api";
+import EventCard from "../EventCard";
+import { CardHeader, CardList, CardTitle } from "./TimelineCard.styled";
 
 export interface TimelineCardProps {
   timeline: Timeline;
   isVisible?: boolean;
-  onShowTimeline?: (timeline: Timeline) => void;
-  onHideTimeline?: (timeline: Timeline) => void;
+  onToggleTimeline?: (timeline: Timeline, isVisible: boolean) => void;
 }
 
 const TimelineCard = ({
   timeline,
   isVisible,
-  onShowTimeline,
-  onHideTimeline,
+  onToggleTimeline,
 }: TimelineCardProps): JSX.Element => {
+  const events = getEvents(timeline.events);
+
   const handleToggle = useCallback(
     (event: ChangeEvent<HTMLInputElement>) => {
-      if (event.target.checked) {
-        onShowTimeline?.(timeline);
-      } else {
-        onHideTimeline?.(timeline);
-      }
+      onToggleTimeline?.(timeline, event.target.checked);
     },
-    [timeline, onShowTimeline, onHideTimeline],
+    [timeline, onToggleTimeline],
   );
 
   return (
@@ -41,9 +39,21 @@ const TimelineCard = ({
       iconVariant="up-down"
       iconPosition="right"
     >
-      <EventList timeline={timeline} />
+      <CardList>
+        {events.map(event => (
+          <EventCard key={event.id} event={event} />
+        ))}
+      </CardList>
     </CollapseSection>
   );
+};
+
+const getEvents = (events: TimelineEvent[] = []) => {
+  return _.chain(events)
+    .filter(e => !e.archived)
+    .sortBy(e => parseTimestamp(e.timestamp))
+    .reverse()
+    .value();
 };
 
 export default memo(TimelineCard);
